% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putio-download.R
\name{bind_putio_download}
\alias{bind_putio_download}
\title{Bind putio download_count to a table with hash}
\usage{
bind_putio_download(tbl)
}
\description{
Bind putio download_count to a table with hash
}
