% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busboyr-package.R
\docType{package}
\name{busboyr-package}
\alias{busboyr}
\alias{busboyr-package}
\title{busboyr: File Searching and Organizing for Put.io}
\description{
Busboy finds links associated with IMDB titles and adds them as a
transfer to put.io It also neatly organizes files on put.io
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/refik/busboyr}
  \item Report bugs at \url{https://github.com/refik/busboyr/issues}
}

}
\author{
\strong{Maintainer}: Refk Turkeli \email{refik.rfk@gmail.com}

}
\keyword{internal}
