% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{get_logger}
\alias{get_logger}
\title{Get a logger function}
\usage{
get_logger(func_name = NULL)
}
\arguments{
\item{func_name}{This value is inferred from the call stack. However,
it is inferred wrong if the function is memoised.}
}
\description{
Get a logger function
}
