% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdb-dataset-query.R
\name{imdb_dataset_episode}
\alias{imdb_dataset_episode}
\title{Get episodes of a series}
\usage{
imdb_dataset_episode(imdb_id, season = NULL, con = db_pool())
}
\description{
Get episodes of a series
}
