% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdb.R
\name{imdb_read}
\alias{imdb_read}
\title{Read imdb dataset file}
\usage{
imdb_read(type = c("episode", "basics"), file_name)
}
\description{
Couldn't solve some weird quoting error and wrote this custom tsv
reading function instead.
}
