% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdb-dataset.R
\name{imdb_read_tsv}
\alias{imdb_read_tsv}
\title{Read imdb dataset file}
\usage{
imdb_read_tsv(type = c("episode", "basics"), file_name)
}
\description{
Couldn't solve some weird quoting error and wrote this custom tsv
reading function instead.
}
