% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-request-organize.R
\name{organize_download}
\alias{organize_download}
\title{Organize the downloaded files}
\usage{
organize_download(putio_user_id, putio_user_download_id)
}
\description{
Organize the downloaded files
}
