% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-request.R
\name{process_season_request}
\alias{process_season_request}
\title{Add a title's season to downloads for the user}
\usage{
process_season_request(season_request_id)
}
\description{
Add a title's season to downloads for the user
}
