% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putio-api.R
\name{putio_account_info}
\alias{putio_account_info}
\title{put.io account information}
\usage{
putio_account_info(user_id = NULL, oauth_token = putio_get_token(user_id))
}
\description{
put.io account information
}
