% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putio-api.R
\name{putio_add_transfer}
\alias{putio_add_transfer}
\title{Add transfer to put.io}
\usage{
putio_add_transfer(user_id = NULL, source, save_parent_id = 0,
  callback_url = NULL, oauth_token = putio_get_token(user_id))
}
\description{
Add transfer to put.io
}
