% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putio-download.R
\name{putio_bind_download_count}
\alias{putio_bind_download_count}
\title{Bind putio download_count to hash}
\usage{
putio_bind_download_count(df_with_hash)
}
\description{
Bind putio download_count to hash
}
