% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putio-cache.R
\name{putio_cached_files_list}
\alias{putio_cached_files_list}
\title{Cached files list response from put.io}
\usage{
putio_cached_files_list(putio_parent_id, con = db_pool())
}
\description{
Cached files list response from put.io
}
