% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putio-cache.R
\name{putio_cached_user}
\alias{putio_cached_user}
\title{Get user information from previous api response}
\usage{
putio_cached_user(putio_user_id)
}
\description{
Get user information from previous api response
}
