% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putio-api.R
\name{putio_create_folder}
\alias{putio_create_folder}
\title{Create folder on put.io}
\usage{
putio_create_folder(user_id = NULL, name, parent_id = 0,
  oauth_token = putio_get_token(user_id))
}
\description{
Create folder on put.io
}
