% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putio-api.R
\name{putio_files_delete}
\alias{putio_files_delete}
\title{Delete  a file on put.io}
\usage{
putio_files_delete(user_id = NULL, file_id,
  oauth_token = putio_get_token(user_id))
}
\description{
Delete  a file on put.io
}
