% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putio-api.R
\name{putio_files_list}
\alias{putio_files_list}
\title{Get users files}
\usage{
putio_files_list(putio_user_id = NULL, parent_id = 0,
  oauth_token = putio_get_token(putio_user_id))
}
\description{
Get users files
}
