% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putio-api.R
\name{putio_files_move}
\alias{putio_files_move}
\title{Rename a file on put.io}
\usage{
putio_files_move(putio_user_id = NULL, user_file_id, parent_id,
  oauth_token = putio_get_token(putio_user_id))
}
\description{
Rename a file on put.io
}
