% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putio-api.R
\name{putio_get_transfer}
\alias{putio_get_transfer}
\title{Get information about a transfer}
\usage{
putio_get_transfer(putio_user_id, putio_user_download_id,
  oauth_token = putio_get_token(putio_user_id))
}
\description{
Get information about a transfer
}
