% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putio-api.R
\name{putio_oauth_redirect_url}
\alias{putio_oauth_redirect_url}
\title{put.io auth url for busboy}
\usage{
putio_oauth_redirect_url(redirect_url = "http://busboy.io")
}
\description{
put.io auth url for busboy
}
