% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdb.R
\name{read_imdb}
\alias{read_imdb}
\title{Read imdb dataset file}
\usage{
read_imdb(type = c("episode", "basics"), file_name)
}
\description{
Couldn't solve some weird quoting error and wrote this custom tsv
reading function instead.
}
