% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-request.R
\name{request_title}
\alias{request_title}
\title{Add titles to download}
\usage{
request_title(putio_user_id, imdb_id, prefer_hd = FALSE,
  season_request_id = NULL)
}
\description{
Add titles to download
}
