% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{start_download}
\alias{start_download}
\title{Start a download for a title}
\usage{
start_download(user_id, request_id, name, source, title_id, season = NULL,
  episode = NULL, season_pack = FALSE)
}
\description{
Start a download for a title
}
