% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title-request.R
\name{title_request_start_download}
\alias{title_request_start_download}
\title{Create title request download}
\usage{
title_request_start_download(putio_user_id, title_request_id, torrent_name,
  torrent_source)
}
\description{
title_request_id can be a vector.
}
