% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torrentapi.R
\name{torrentapi_imdb_search}
\alias{torrentapi_imdb_search}
\title{Search torrents with imdb id}
\usage{
torrentapi_imdb_search(imdb_id, season = NULL, episode = NULL,
  season_pack_only = FALSE)
}
\description{
Search torrents with imdb id
}
