% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer-complete.R
\name{transfer_complete_consume}
\alias{transfer_complete_consume}
\title{Consumes transfer complete tasks from putio}
\usage{
transfer_complete_consume()
}
\description{
Consumes transfer complete tasks from putio
}
