% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer-complete.R
\name{transfer_complete_refresh_shiny}
\alias{transfer_complete_refresh_shiny}
\title{Notify user about updates}
\usage{
transfer_complete_refresh_shiny(putio_user_id)
}
\description{
Notify user about updates
}
