% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putio-api.R
\name{putio_files_rename}
\alias{putio_files_rename}
\title{Rename a file on put.io}
\usage{
putio_files_rename(user_id = NULL, file_id, new_name,
  oauth_token = putio_get_token(user_id))
}
\description{
Rename a file on put.io
}
